supervision-functions: 4.0 version bump
Merged supervision set up and clean up shell functions to runscript API;
meaning that pre() is renamed to start_pre(), post() is renamed stop_post();
and log_pre() to log_start_pre(), log_post() to log_stop_post(). This API
change was necessary for ease of service creation and maintainability to
not have to split service for setup and clean up.